// frontend/src/api/generated/models/generateReviewedQuestionsDtoDifficulty.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * Lammah Game API
 * Runtime HTTP contract for Lammah game administration, content, and gameplay.
 * OpenAPI spec version: 1.0.0
 */

export type GenerateReviewedQuestionsDtoDifficulty =
  (typeof GenerateReviewedQuestionsDtoDifficulty)[keyof typeof GenerateReviewedQuestionsDtoDifficulty];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const GenerateReviewedQuestionsDtoDifficulty = {
  easy: "easy",
  medium: "medium",
  hard: "hard",
} as const;